pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "zip-to-github-uploader"
version = "1.0.0"
description = "Intelligentes Werkzeug zur Automatisierung von ZIP-Projekt-Uploads in GitHub-Repositories"
readme = "README.md"
authors = [{ name = "swisscomfort", email = "hello@swisscomfort.dev" }]
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "altair>=5.0.0",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "pylint>=2.17",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = ["sphinx>=5.0", "sphinx-rtd-theme>=1.0"]

[project.urls]
Homepage = "https://github.com/swisscomfort/zip-to-github-uploader"
Documentation = "https://github.com/swisscomfort/zip-to-github-uploader/wiki"
Repository = "https://github.com/swisscomfort/zip-to-github-uploader.git"
Issues = "https://github.com/swisscomfort/zip-to-github-uploader/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "C0111", # missing-docstring
    "C0103", # invalid-name
]

[tool.pylint.FORMAT]
max-line-length = 100
